refactor(engine): main file
