+++ packages/engine/package.json
@@ -49,12 +49,6 @@
   "bugs": {
     "url": "https://github.com/Alwatr/store/issues"
   },
-  "nano-build": {
-    "entryPoints": [
-      "src/main.ts",
-      "src/provider.ts"
-    ]
-  },
   "prettier": "@alwatr/prettier-config",
   "scripts": {
     "b": "yarn run build",
@@ -81,7 +75,6 @@
     "@alwatr/tsconfig-base": "^1.0.5",
     "@alwatr/type": "^1.2.4",
     "@types/node": "^20.10.5",
-    "esbuild": "^0.19.10",
     "npm-run-all": "^4.1.5",
     "typescript": "^5.3.3"
   }
